디렉토리 구조 변화 Lotto 폴더를 만들어 관련 클래스 코드 이동
